+++ scripts/main_overparameterized_time.py
@@ -21,12 +21,13 @@ if __name__ == "__main__":
     ]
     # Check if there is a process running with name containing 'python'
     for process in psutil.process_iter(["name"]):
-        if "python" in process.info["name"]:
-            cmdline = process.cmdline()
-            if len(cmdline) > 1:
-                if cmdline[1] in files:
-                    print(f"Killing process {process.pid}")
-                    process.kill()
+        if process.info["name"] is not None:
+            if "python" in process.info["name"]:
+                cmdline = process.cmdline()
+                if len(cmdline) > 1:
+                    if cmdline[1] in files:
+                        print(f"Killing process {process.pid}")
+                        process.kill()
 
     prcesses = dict()
